all the initial expected micro services has been added.

+++ Ticket-Katum-gateway/bff/web-bff/src/main/java/com/ticketkatum/dto/hotel/HotelRecommendation.java
@@ -22,7 +22,7 @@ public class HotelRecommendation {
     private BigDecimal maxPrice;
 
     private Double averageRating;
-    private Integer totalReviews;
+    private Double totalReviews;
 
     // distance from user (in KM)
     private Double distance;
